refactor: extract pure statistics into domain/stats module
