+++ origa_ui/src/pages/home/dashboard_stats.rs
@@ -10,24 +10,6 @@ const EN_MONTHS_SHORT: [&str; 12] = [
     "Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec",
 ];
 
-#[derive(Clone, Default)]
-pub struct TodayOverview {
-    pub new_count: usize,
-    pub learned_count: usize,
-    pub in_progress_count: usize,
-    pub difficult_count: usize,
-    pub new_delta: Option<i32>,
-    pub learned_delta: Option<i32>,
-    pub in_progress_delta: Option<i32>,
-    pub difficult_delta: Option<i32>,
-}
-
-impl TodayOverview {
-    pub fn total(&self) -> usize {
-        self.new_count + self.learned_count + self.in_progress_count + self.difficult_count
-    }
-}
-
 #[derive(Clone)]
 pub struct RecentlyStudiedItem {
     pub card_id: String,
@@ -45,53 +27,6 @@ pub struct ActivityDataPoint {
     pub in_progress: f64,
     pub new_count: f64,
     pub difficult: f64,
-}
-
-pub fn compute_today_overview(
-    knowledge_set: &KnowledgeSet,
-    history: &[DailyHistoryItem],
-) -> TodayOverview {
-    let mut overview = TodayOverview::default();
-
-    for study_card in knowledge_set.study_cards().values() {
-        let card = study_card.card();
-        if matches!(CardType::from(card), CardType::Phrase) {
-            continue;
-        }
-        let memory = study_card.memory();
-        if memory.is_new() {
-            overview.new_count += 1;
-        } else if memory.is_high_difficulty() {
-            overview.difficult_count += 1;
-        } else if memory.is_known_card() {
-            overview.learned_count += 1;
-        } else {
-            overview.in_progress_count += 1;
-        }
-    }
-
-    if history.len() >= 2 {
-        let mut sorted: Vec<_> = history.to_vec();
-        sorted.sort_by_key(|a| a.timestamp());
-        let today = sorted.last().expect("last item exists");
-        let yesterday = sorted
-            .iter()
-            .rev()
-            .nth(1)
-            .expect("second-to-last item exists");
-
-        overview.new_delta = Some((today.new_words() as i64 - yesterday.new_words() as i64) as i32);
-        overview.learned_delta =
-            Some((today.known_words() as i64 - yesterday.known_words() as i64) as i32);
-        overview.in_progress_delta =
-            Some((today.in_progress_words() as i64 - yesterday.in_progress_words() as i64) as i32);
-        overview.difficult_delta = Some(
-            (today.high_difficulty_words() as i64 - yesterday.high_difficulty_words() as i64)
-                as i32,
-        );
-    }
-
-    overview
 }
 
 pub fn compute_studied_today(
@@ -209,36 +144,6 @@ pub fn compute_30day_chart_data(
             }
         })
         .collect()
-}
-
-#[derive(Clone, Default)]
-pub struct RatingRatio {
-    pub percentage: u8,
-    pub positive_count: usize,
-    pub negative_count: usize,
-}
-
-pub fn compute_rating_ratio(history: &[DailyHistoryItem]) -> Option<RatingRatio> {
-    let mut positive = 0usize;
-    let mut negative = 0usize;
-
-    for item in history {
-        positive += item.positive_ratings();
-        negative += item.negative_ratings();
-    }
-
-    let total = positive + negative;
-    if total == 0 {
-        return None;
-    }
-
-    let percentage = (positive as f64 / total as f64 * 100.0).round() as u8;
-
-    Some(RatingRatio {
-        percentage,
-        positive_count: positive,
-        negative_count: negative,
-    })
 }
 
 #[derive(Clone, Default)]
